Cleanup code for pdf scraper
- Move the download pdf method to base pdf scraper
